+++ lib/gestures/gestures.dart
@@ -1,0 +1,60 @@
+import 'package:flutter/material.dart';
+
+class SampleGestures extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Touch ripples'),
+      ),
+      body: Center(
+        child: HandleTaps(),
+      ),
+    );
+  }
+}
+
+// https://flutter.dev/docs/cookbook/gestures/handling-taps
+class HandleTaps extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    // The GestureDetector wraps the button.
+    return GestureDetector(
+      // When the child is tapped, show a snackbar.
+      onTap: () {
+        final snackBar = SnackBar(content: Text("Tap"));
+
+        ScaffoldMessenger.of(context).showSnackBar(snackBar);
+      },
+      // The custom button
+      child: Container(
+        padding: EdgeInsets.all(12.0),
+        decoration: BoxDecoration(
+          color: Theme.of(context).buttonColor,
+          borderRadius: BorderRadius.circular(8.0),
+        ),
+        child: Text('My Button'),
+      ),
+    );
+  }
+}
+
+// https://flutter.dev/docs/cookbook/gestures/ripples
+class TouchRipples extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    // The InkWell wraps the custom flat button widget.
+    return InkWell(
+      // When the user taps the button, show a snackbar.
+      onTap: () {
+        ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+          content: Text('Tap'),
+        ));
+      },
+      child: Container(
+        padding: EdgeInsets.all(12.0),
+        child: Text('Flat Button'),
+      ),
+    );
+  }
+}
